requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cryptography==41.0.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langchain-openai==0.0.3
langchainhub==0.1.14
langsmith==0.0.83
llama-hub==0.0.75
llama-index==0.9.36
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.9.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==3.11.0
plotly==5.21.0
protobuf==4.25.2
psutil==5.9.8
pyaml==23.12.0
pyarrow==15.0.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
retrying==1.3.4
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.51
streamlit==1.30.0
streamlit-ace==0.1.1
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
types-requests==2.31.0.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
waii-sdk-py==1.13.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
pyjwt~=2.8.0
